finish working on interfaces

+++ appname/interfaces/model_controller.py
@@ -1,4 +1,4 @@
-from appname.src.interfaces.model import Model
+from appname.interfaces.model import Model
 
 class ModelController():
 
@@ -14,16 +14,20 @@ class ModelController():
         
         raise ValueError("Model validation failed.")
 
-    
     def get(self, **filters):
+        if not filters:
+            raise ValueError("Filters cannot be empty for get operation.")
         return self.model_class.get(**filters)
     
     def get_all(self, **filters):
         return self.model_class.get_all(**filters)
     
     def delete(self, model):
+        if not isinstance(model, self.model_class):
+            raise TypeError("Invalid model instance for deletion.")
         return model.delete()
     
     def delete_all(self, **filters):
+        if not filters:
+            raise ValueError("Filters cannot be empty for delete_all operation.")
         return self.model_class.delete_all(**filters)
-
